streamlit_vendeur/requirements.txt
streamlit
asyncio
nest-asyncio
pandas
numpy
scikit-learn
transformers
torch
pathlib
joblib
folium
streamlit-folium
matplotlib
seaborn
geopy
python-Levenshtein
python-dotenv
plotly
openai
requests
pillow
